+++ test_script.kt
@@ -1,9 +1,0 @@
-import com.ikolvi.tracelet.sdk.model.TraceletConfig
-
-fun main() {
-    val map = mapOf(
-        "speedStationaryDelay" to 2
-    )
-    val conf = TraceletConfig.fromMap(map)
-    println(conf.motion.speedStationaryDelay)
-}