+++ packages/tracelet/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.2.0
 
- - **FEAT**: Add reverse geocoding functionality. ([0fe7b89a](https://github.com/Ikolvi/Tracelet/commit/0fe7b89aad0e22ea28cf81dd81723a534300c175))
+- **FEAT**: Introduced new `tracelet_sync` package for offline SQLite persistence and automatic HTTP synchronization.
+- **FEAT**: Add reverse geocoding (`resolveAddress`) functionality for automatic address lookups.
 
 ## 3.1.14